pyproject.toml
[project]
readme = "README.md"
name = "sorrel"
version = "0.1.0"
description = "A flexible framework for developing reinforcement learning algorithms."
license = "MIT"
authors = [
    {name = "Yibing Ju", email = "bing.ju@mail.utoronto.ca"}, 
    {name = "Rebekah Gelpi", email = "rebekah.gelpi@mail.utoronto.ca"},
    {name = "Shon Verch", email = "verchshon@gmail.com"},
    {name = "Ethan Jackson", email = "jackson.ethan.c@gmail.com"},
]

[tool.poetry]
packages = [{ include = "sorrel" }, { include = "examples" }]

[tool.poetry.dependencies]
python = "^3.11.0"
matplotlib = "^3.6.1"
numpy = "^1.23.4"
pyyaml = "^6.0.2"

torch = "^2.5.1"
jax = "^0.4.35"
flax = "^0.10.1"

# used in buffer.py
dill = "^0.3.9"
# used in human_player.py
ipython = "^8.29.0"
# used in utils.py
pillow = "^11.0.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"

# for docs
sphinx = "^8.1.3"
sphinx-rtd-theme = "^3.0.2"
myst-nb = "^1.1.2"
sphinx-autodoc2 = "^0.5.0"

[tool.poetry.group.extras]
optional = true

[tool.poetry.group.extras.dependencies]
tensorboard = "^2.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
